# Add `ExLlamaV2Sampler.Settings.logits_processor`