--- a/exllamav2/generator/sampler.py
+++ b/exllamav2/generator/sampler.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 from dataclasses import dataclass, field
+from typing import Optional, Callable
 import torch
 import torch.nn.functional as F
 from exllamav2 import ExLlamaV2Tokenizer
@@ -70,6 +71,8 @@ class ExLlamaV2Sampler:
         tfs: float = 0
         typical: float = 0
         skew: float = 0
+
+        logits_processor: Optional[Callable] = None
 
         temperature_last: bool = False
 
@@ -269,6 +272,7 @@ class ExLlamaV2Sampler:
     def sample(
         logits: torch.tensor,
         settings: Settings,
+        input_ids: torch.tensor,
         sequence_ids: torch.tensor,
         random: float,
         tokenizer: ExLlamaV2Tokenizer,
@@ -288,6 +292,9 @@ class ExLlamaV2Sampler:
 
         :param settings:
             ExLlamaV2Sampler.Settings
+
+        :param input_ids:
+            The prompt portion of sequence_ids, shape (batch_size, seq_len)
 
         :param sequence_ids:
             Past token IDs to consider for repetition penalty etc., shape (batch_size, seq_len)
@@ -353,6 +360,12 @@ class ExLlamaV2Sampler:
             logits = settings.cfg_scale * logits[0] + (1 - settings.cfg_scale) * logits[1]
             logits = logits.unsqueeze(0)
             batch_size = 1
+
+        # Apply logits processor
+
+        if settings.logits_processor:
+            generated_ids = sequence_ids[:, input_ids.shape[1]:]
+            logits = settings.logits_processor(generated_ids, logits)
 
         # Prepare filter