# Overview / Motivation
Implements `ExLlamaV2Sampler.Settings.logits_processor` which allows us to
- Take advantage of third party libraries logits processors, such as Outlines which implements JSON Schema, regex, and Lark structured generation logits processors
- Enable implementation ofnew sampling techniques through torch-based logits processors
- Have a path towards faster sampling through keeping `logits` on GPU and using torch, rather than sampling on CPU.

## Changes
- Introduce `ExLlamaV2Sampler.Settings.logits_processor` which allows for logits filtering and augmentation with torch
  - Make necessary changes to generation-related code to ensure logits processor is applied
- Introduce `tests/test_logits_processors.py` which is the same as `tests/test.py` but using a logits processor for all sampling
- Introduce `examples/json_schema_outlines.py`

## Performance
- No change in performance for `tests.py` between this branch and `master`
- Slower when optional `logits_processor` argument is enabled
  - This builds a path for a future change where, ExLlamaV2 sampling is even **faster** by keeping `logits` and `input_ids` on GPU and using torch (out of scope for this PR


| Test                                    | normal | streaming (prompt) | streaming (response) | batched |
|-----------------------------------------|-------------------:|------------------------------:|---------------------------------:|--------------------:|
| master -> tests.py                      |             157.71 |                      3491.57  |                          179.95 |               10.14 |
| this branch -> tests.py                 |             158.39 |                      3559.52  |                          178.04 |               10.15 |
| this branch -> test_logits_processor.py |             122.66 |                      3576.11  |                          134.56 |                9.97 |



<details>

`master` -> `tests.py`
```
Generating, normal
...
Response generated in 1.51 seconds, 150 tokens, 99.38 tokens/second

Generating, streaming
...
Prompt processed in 0.00 seconds, 15 tokens, 3693.04 tokens/second
Response generated in 1.10 seconds, 150 tokens, 136.67 tokens/second

Generating, batched
...
Response generated in 17.87 seconds, 40 tokens, throughput 8.95 tokens/second
```
(Note: `Generating, batched multi cache` fails in `master`, not due to this PR)

`sampler-logits-processor` -> `tests.py`

```
Generating, normal
...
Response generated in 1.51 seconds, 150 tokens, 99.46 tokens/second

Generating, streaming
...
Prompt processed in 0.00 seconds, 15 tokens, 3534.72 tokens/second                                                            
Response generated in 1.12 seconds, 150 tokens, 133.63 tokens/second                                                          

Generating, batched
...
Response generated in 18.26 seconds, 40 tokens, throughput 8.76 tokens/second
```

`sampler-logits-processor` -> `test_logits_processor.py`
```
Generating, normal
...
Response generated in 1.22 seconds, 150 tokens, 122.66 tokens/second

Generating, streaming
...
Prompt processed in 0.00 seconds, 15 tokens, 3576.11 tokens/second
Response generated in 1.11 seconds, 150 tokens, 134.56 tokens/second

Generating, batched
...
Response generated in 16.04 seconds, 40 tokens, throughput 9.97 tokens/second
```

</details>

# Tests

All tests pass except for `tests/test.py` / `Generating, batched multi cache` which also fails in `master` 

<details>

```
Generating, batched multi cache

Traceback (most recent call last):
  File "/root/exllamav2/tests/test.py", line 249, in <module>
    tests(q_model_directory, False, False)
  File "/root/exllamav2/tests/test.py", line 241, in tests
    if model.is_quant(): test_multicache(40)
  File "/root/exllamav2/tests/test.py", line 211, in test_multicache
    logits = model.forward(inputs, caches, input_mask = None).float().cpu()
  File "/opt/conda/lib/python3.10/site-packages/torch/utils/_contextlib.py", line 116, in decorate_context
    return func(*args, **kwargs)
  File "/root/exllamav2/exllamav2/model.py", line 809, in forward
    result = self.forward_chunk(input_ids = input_ids,
  File "/opt/conda/lib/python3.10/site-packages/torch/utils/_contextlib.py", line 116, in decorate_context
    return func(*args, **kwargs)
  File "/root/exllamav2/exllamav2/model.py", line 922, in forward_chunk
    past_len = cache.current_seq_len
AttributeError: 'list' object has no attribute 'current_seq_len'
```

</details>